[MOV] point_of_sale,*: split `pos_order_line` from `pos_order`
*: pos_mrp, pos_sale, pos_self_order

In this commit:
------------------
- We have separated the `pos_order_line` model from the `pos_order` model to
better organize the files and improve the overall code structure, making the
codebase easier to understand, maintain, and navigate.

task: 5473161
