+++ addons/point_of_sale/models/__init__.py
@@ -16,6 +16,7 @@ from . import digest
 from . import pos_category
 from . import pos_config
 from . import pos_order
+from . import pos_order_line
 from . import pos_session
 from . import pos_snooze
 from . import product_pricelist
