+++ addons/pos_sale/models/__init__.py
@@ -4,6 +4,7 @@
 from . import account_move
 from . import pos_config
 from . import pos_order
+from . import pos_order_line
 from . import product_template
 from . import crm_team
 from . import pos_session
