+++ addons/pos_self_order/models/__init__.py
@@ -4,6 +4,7 @@ from . import ir_http
 from . import pos_category
 from . import pos_config
 from . import pos_order
+from . import pos_order_line
 from . import pos_restaurant
 from . import pos_payment_method
 from . import pos_self_order_custom_link
