+++ addons/pos_self_order/models/pos_order.py
@@ -8,32 +8,6 @@ from odoo.exceptions import UserError
 _logger = logging.getLogger(__name__)
 
 
-class PosOrderLine(models.Model):
-    _inherit = "pos.order.line"
-
-    combo_id = fields.Many2one('product.combo', string='Combo reference')
-
-    @api.model_create_multi
-    def create(self, vals_list):
-        for vals in vals_list:
-            if (vals.get('combo_parent_uuid')):
-                vals.update([
-                    ('combo_parent_id', self.search([('uuid', '=', vals.get('combo_parent_uuid'))]).id)
-                ])
-            if 'combo_parent_uuid' in vals:
-                del vals['combo_parent_uuid']
-        return super().create(vals_list)
-
-    def write(self, vals):
-        if (vals.get('combo_parent_uuid')):
-            vals.update([
-                ('combo_parent_id', self.search([('uuid', '=', vals.get('combo_parent_uuid'))]).id)
-            ])
-        if 'combo_parent_uuid' in vals:
-            del vals['combo_parent_uuid']
-        return super().write(vals)
-
-
 class PosOrder(models.Model):
     _inherit = "pos.order"
